LocalMaxMatcher now accepts list of vertices that should not be matched to each other some bugfixes

+++ networkit/cpp/structures/Partition.cpp
@@ -193,7 +193,7 @@ edgeweight Partition::calculateCutWeight(const Graph& G) {
 
 double Partition::getImbalance(count desiredPartitions) {
 	double maxImbalance = 0;
-	double averageSize = numberOfElements() / desiredPartitions;
+	double averageSize = double(numberOfElements()) / desiredPartitions;
 	for (auto entry : subsetSizeMap()) {
 		double imbalance = std::abs(entry.second - averageSize) / averageSize;
 		if (maxImbalance < imbalance) maxImbalance = imbalance;
